Renamed planners removed test_
@amdshameer. I have removed the test_ infront of the planner nodes.

Please only work on these files and do not check in test files unless they are real unit tests.
I believe these were just you testing if it worked.
Use Git to test. create a new branch, work on the original files in the new branch and if it doesn't work, then delete the branch and checkout the master again.

This is related to comments on issue mas-group/sdp_ws2013_ompl#51